Rearrange imports to something that makes a little more sense
