+++ src/metadata.js
@@ -1,11 +1,7 @@
-class Metadata {
+export default class Metadata {
   constructor(name, debug = false, testing = false) {
     this.name = name;
     this.debug = debug;
     this.testing = testing;
   }
 }
-
-module.exports = {
-  Metadata,
-};
